Use float64

+++ pkg/price/price.go
@@ -4,6 +4,6 @@ func VisualPriceToDBPrice(price float32) uint64 {
 	return uint64(price * 1000000000000)
 }
 
-func DBPriceToVisualPrice(price uint64) float32 {
-	return float32(price / 1000000000000.0)
+func DBPriceToVisualPrice(price uint64) float64 {
+	return float64(price / 1000000000000.0)
 }
